+++ lib/const/colors.dart
@@ -1,9 +1,10 @@
 import 'package:flutter/material.dart';
 
 class AppColors {
-  final dynamic backgroundColor = const Color.fromARGB(255, 255, 255, 255);
-  final dynamic boxColor=  const Color.fromARGB(255, 132, 81, 207);
-  final dynamic textColor= const Color(0xFF111417);
-  final dynamic smallText = const Color(0xFF647487);
-  final dynamic hyperlink = const Color(0xFF3889E5);
+  static final backgroundColor = const Color.fromARGB(255, 255, 255, 255);
+  static final boxColor=  const Color.fromARGB(255, 132, 81, 207);
+  static final textColor= const Color(0xFF111417);
+  static final smallText = const Color(0xFF647487);
+  static final hyperlink = const Color(0xFF3889E5);
+  static final barColor = const Color(0xFFB9A3F4);
 }