+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Commit Changelog
 - Added:
+    - User icon in the dashboard app bar (app icon for now)
+    - 'Cairo' font family in the 'fonts' folder
+    - TaskModel class to manage tasks efficiently (will come into play once we implement the backend)
+    - Non-Functional Bottom Nav Bar
 - Fixed:
-    - Solved an issue where app icon is displayed black!
+    - Changed 'AppColor' class attributes to static to not create an instance of the class each time
 - Updated:
+    - Home Page with ugly design for now until backend is implemented